djangorest/requirements.txt
Django==2.0.7
django-rest-framework==0.1.0
djangorestframework==3.8.2
whitenoise==2.0.0
gunicorn==19.9.0
django-cors-headers==2.4.0
